+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 ### Fixed
 - (`sklearn`) Crash when there were no informative trees in predictors.
 
+### Deprecated
+- `core`:
+  - The `build_multi_table_dictionary_domain` helper function.
+
 ## 10.3.0.0 - 2025-02-10
 
 ### Fixed